Add shield_functions pass to improve deduplication
By wrapping top-level function definitions in one additional layer of indirection, we reduce the number of opaque function representations generated after closure representation.  This works because wrapping a function in another function "shields" the inner function's representation from having to potentially unify with other representations, and allows it to instead become an interenal representation variable from the perspective of the outer function.