Add data layout field to generated LLVM
This fixes the valgrind errors we were seeing after enabling optimizations.  If we don't include the data layout field, then LLVM's middle end optimizer guesses (!) target-specific sizes and alignments incorrectly.  Notably, this occurs when we use a constant 'getelementptr' expression to compute the size of a type, as when generating calls to 'malloc'.