Annotate expressions with ids and fates
This will be useful for attaching per-version RC information to expressions in later RC elision passes.

We don't strictly need to store expression fates in a separate data structure keyed by ids; we could just as well store the fates inline.  However, this is more consistent with how we store per-occurrence fate information, and it's a nice demonstration / test use case for the more general expression id system.