Enforce that RC ops never touch mutated arrays
This incidentally fixes multiple soundness bugs that existed in the previous implementation.  In mutation analysis, this commit fixes an issue where mutations in array ops were not correctly propagated across internal aliases within the arguments to the op.  In RC elision, this fixes the soundness of 'get', which I believe (although I have not confirmed) was vulnerable to a use-after-free issue if the parent array of an item acquired with 'get' was mutated (via a separately-owned reference) before the lifetime of the borrowed reference to the item ended.