Add 'words_trie' benchmark
Programs implementing this benchmark take as input (1) a newline-delimited list of words, and (2) a newline-delimited list of "queries."  They return a list of integers, one for each query, reporting how many times each query occurred among the original list of words.  We use a nybble trie to track the count of each word.